pyproject.toml
[project]
name = "uba-ceia-proy-final"
version = "0.1.0"
description = "Dependencias del proyecto final"
authors = [{ name = "brunomaso1", email = "brunomaso1@hotmail.com" }]
license = { text = "CC" }
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pymongo (>=4.9.2,<4.10.0)",
    "boto3 (>=1.38.4,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "cvat-sdk (==2.32.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pycocotools (>=2.0.8,<3.0.0)",
    "layoutparser (>=0.3.4,<0.4.0)",
    "fastkml (>=1.1.0,<2.0.0)",
    "kml2geojson (>=5.1.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "geopandas (>=1.0.1,<2.0.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "gputil (>=1.4.0,<2.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "torchaudio (>=2.7.0,<3.0.0)",
    "ultralytics (>=8.3.139,<9.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "fiftyone (==1.2.0)",
]

[tool.poetry]
package-mode = false

# packages = [
#     # Incluye el directorio "modulo-mini-apps" como paquete
#     { include = "modulo-mini-apps", from = "ceia-proyecto-final" },
# ] # TODO: Chequear si funciona esto.

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu126"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-gpu"}
torchvision = {source = "pytorch-gpu"}
torchaudio = {source = "pytorch-gpu"}
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
